xclhal2_mem to xrt_mem
