+++ src/CMake/dkms.cmake
@@ -156,6 +156,7 @@ SET (XRT_DKMS_CORE_INCLUDES
   include/xclbin.h
   include/xclerr.h
   include/xclhal2_mem.h
+  include/xrt_mem.h
   )
 
 SET (XRT_DKMS_COMMON_XRT_DRV
