+++ src/runtime_src/core/include/CMakeLists.txt
@@ -11,7 +11,8 @@ set(XRT_HEADER_SRC
   xcl_macros.h
   xcl_app_debug.h
   xclperf.h
-  xclhal2_mem.h)
+  xclhal2_mem.h
+  xrt_mem.h)
 
 install (FILES ${XRT_HEADER_SRC} DESTINATION ${XRT_INSTALL_INCLUDE_DIR})
