+++ src/runtime_src/core/include/xclhal2_mem.h
@@ -18,37 +18,6 @@
 #ifndef _XCLHAL2_MEM_H_
 #define _XCLHAL2_MEM_H_
 
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/**
- * XCL BO Flags bits layout
- *
- * bits  0 ~ 15: DDR BANK index
- * bits 16 ~ 31: BO flags
- */
-#define	XCL_BO_FLAGS_CACHEABLE		(1 << 26)
-#define	XCL_BO_FLAGS_SVM		(1 << 27)
-#define	XCL_BO_FLAGS_DEV_ONLY		(1 << 28)
-#define	XCL_BO_FLAGS_HOST_ONLY		(1 << 29)
-#define	XCL_BO_FLAGS_P2P		(1 << 30)
-#define	XCL_BO_FLAGS_EXECBUF		(1 << 31)
-
-/**
- * This is the legacy usage of XCL DDR Flags.
- *
- * byte-0 lower 4 bits for DDR Flags are one-hot encoded
- */
-enum xclDDRFlags {
-    XCL_DEVICE_RAM_BANK0 = 0x00000000,
-    XCL_DEVICE_RAM_BANK1 = 0x00000002,
-    XCL_DEVICE_RAM_BANK2 = 0x00000004,
-    XCL_DEVICE_RAM_BANK3 = 0x00000008,
-};
-
-#ifdef __cplusplus
-}
-#endif
+#include "xrt_mem.h"
 
 #endif
